Add files via upload
Here Are the questions that are solved in the PowerBi Reports
Question 1:
The time series of the revenue data for the year 2011 only, looking into revenue for each month including Seasonal trends and their reason.
Question 2:
The top 10 countries which are generating the highest revenue. Additionally,  Quantity sold along with the revenue generated, Excluding United Kingdom.
Question 3:
The top 10 customers by revenue. Create a visual that shows the greatest revenue generating customer at the start and gradually declines to the lower revenue generating customers.
Question 4:
Gain insights on the demand for the products in Various Regions excluding United Kingdom.